Move geodesic to point aligner within metric
