+++ geomstats/learning/aac.py
@@ -1,8 +1,6 @@
 import logging
 import random
 
-import scipy
-
 import geomstats.backend as gs
 from geomstats.learning._sklearn_wrapper import WrappedLinearRegression, WrappedPCA
 from geomstats.learning.frechet_mean import FrechetMean
@@ -75,7 +73,6 @@ class AACGPC:
         self.init_point = init_point
 
         self.pca_solver = WrappedPCA(n_components=n_components)
-        self.aligner = None
 
     @property
     def components_(self):
@@ -97,15 +94,6 @@ class AACGPC:
     def mean_(self):
         return self.pca_solver.reshaped_mean_
 
-    def set_default_aligner(self, s_min, s_max, n_sample_points=10):
-        self.set_aligner("default", s_min, s_max, n_sample_points=n_sample_points)
-
-    def set_aligner(self, aligner, *args, **kwargs):
-        if aligner == "default":
-            self.aligner = PointToGeodesicAligner(self.metric, *args, **kwargs)
-        else:
-            self.aligner = aligner
-
     def fit(self, X, y=None):
         x = random.choice(X) if self.init_point is None else self.init_point
         aligned_X = self.metric.align_point_to_point(x, X)
@@ -124,7 +112,7 @@ class AACGPC:
                 initial_point=mean, initial_tangent_vec=direc
             )
 
-            aligned_X = self.aligner.align(geodesic, aligned_X)
+            aligned_X = self.metric.align_point_to_geodesic(geodesic, aligned_X)
             self.pca_solver.fit(aligned_X)
             expl_ = self.pca_solver.explained_variance_ratio_[0]
 
@@ -134,95 +122,6 @@ class AACGPC:
         _warn_max_iterations(iteration, self.max_iter)
 
         return self
-
-
-class PointToGeodesicAligner:
-    # TODO: move to metric
-    # TODO: create base class
-
-    def __init__(self, metric, s_min, s_max, n_sample_points=10):
-        self.metric = metric
-        self.s_min = s_min
-        self.s_max = s_max
-        self.n_sample_points = n_sample_points
-
-        self.min_dists_ = None
-
-    @property
-    def _s(self):
-        return gs.linspace(self.s_min, self.s_max, num=self.n_sample_points)
-
-    def _get_gamma_s(self, geodesic):
-        return geodesic(self._s)
-
-    def align(self, geodesic, x):
-        gamma_s = self._get_gamma_s(geodesic)
-
-        n_points = 1 if gs.ndim(x) == 2 else gs.shape(x)[0]
-        if n_points > 1:
-            gamma_s = gs.repeat(gamma_s, n_points, axis=0)
-            rep_x = gs.concatenate([x for _ in range(self.n_sample_points)])
-        else:
-            rep_x = x
-
-        dists = gs.reshape(
-            self.metric.dist(gamma_s, rep_x), (self.n_sample_points, n_points)
-        )
-
-        min_dists_idx = gs.argmin(dists, axis=0)
-        perm_indices = min_dists_idx * n_points + gs.arange(n_points)
-        if n_points == 1:
-            perm_indices = perm_indices[0]
-
-        perms = gs.take(self.metric.perm_, perm_indices, axis=0)
-
-        # TODO: delete?
-        self.min_dists_ = gs.take(
-            gs.transpose(dists),
-            min_dists_idx + gs.arange(n_points) * self.n_sample_points,
-        )
-
-        return self.metric.space.permute(x, perms)
-
-
-class GeodesicToPointAligner:
-    def __init__(self, metric, method="BFGS"):
-        self.metric = metric
-
-        self.method = method
-
-        self.opt_results_ = None
-
-    def _objective(self, s, x, geodesic):
-        point = geodesic(s)
-        dist = self.metric.dist(point, x)
-
-        return dist
-
-    def align(self, geodesic, x):
-        n_points = 1 if gs.ndim(x) == 2 else gs.shape(x)[0]
-
-        if n_points == 1:
-            x = gs.expand_dims(x, axis=0)
-
-        perms = []
-        min_dists = []
-        opt_results = []
-        for xx in x:
-            s0 = 0.0
-            res = scipy.optimize.minimize(
-                self._objective, x0=s0, args=(xx, geodesic), method=self.method
-            )
-            perms.append(self.metric.perm_[0])
-            min_dists.append(res.fun)
-
-            opt_results.append(res)
-
-        self.min_dists_ = gs.array(min_dists)
-        self.opt_results_ = opt_results
-
-        new_x = self.metric.space.permute(x, gs.array(perms))
-        return new_x[0] if n_points == 1 else new_x
 
 
 class AACRegression:
